web_ui/requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.2
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1
pdfplumber>=0.10.0
pdf2image>=1.16.0
tqdm>=4.65.0
requests>=2.30.0
watchdog>=3.0.0
Pillow>=10.0.0
